Refactor currency storage

+++ MixinServices/MixinServices/Services/API/Model/FiatMoney.swift
@@ -4,6 +4,9 @@ import WCDBSwift
 public struct FiatMoney: Codable {
     
     public let code: String
-    public let rate: Double
+    
+    // Decoding decimal numbers directly from ASCII string is supported by JSONDecoder since Swift 5.0
+    // See https://github.com/apple/swift-corelibs-foundation/pull/1657/files
+    public let rate: Decimal
     
 }
